move out of utility and into app/tool/main
I don't think it'll be used anywhere else and it saves recompilation
time if I don't change a header that is used a lot of places
