+++ include/pflib/utility.h
@@ -5,15 +5,6 @@
 #include <string>
 #include <cstdint>
 #include <span>
-#include <iostream>
-#include <algorithm>
-#include <filesystem>
-#include <cstdlib>
-#include <cstdio>
-#include <memory>
-#include <stdexcept>
-#include <string>
-#include <array>
 
 namespace pflib {
 
@@ -48,32 +39,5 @@ bool endsWith(const std::string& full, const std::string& ending);
  * @return value of CRC
  */
 uint32_t crc(std::span<uint32_t> data);
-
-/**
- * Execute a command and capture its output into a string
- *
- * The maximum output is 128 characters.
- *
- * @param[in] cmd command C-string to run
- * @return up to 128 characters of output by cmd
- */
-std::string exec(const char* cmd);
-
-/**
- * Check if the firmware supporting the HGCROC is active
- *
- * @return true if correct firmware is active
- */
-bool is_fw_active();
-
-/**
- * Get the current firmware version
- *
- * This takes a moment because we simply retrieve the version
- * using an rpm query and the exec function.
- *
- * @return string holding the full firmware version
- */
-std::string fw_version();
 
 }  // namespace pflib
